AI/requirements.txt
click
flask
gunicorn
itsdangerous
jinja2
markupSafe
werkzeug
pillow
requests
tensorflow
scikit-learn
matplotlib